Diseño de las ventanas y insert de apuestas para comprobar metodos

+++ RetoFinal/BDA/RetoFinal-BDA/APUESTAS DEPORTIVAS-CreacionTablas.sql
@@ -30,7 +30,7 @@ create table Usuario (
 create table Apuesta (
 	Cod_Apuesta char(3) primary key,
     Fecha_Apuesta date not null,
-    Cuota int not null,
+    Cuota float not null,
     constraint check_Cuota check (Cuota>1));
     
 #creacion de tabla gestionar
